This implements the first phase of true Bluetooth communication. A number of things have been done:
1) The "dummy" data has been removed from the Central displays. You will only see a Peripheral if one is detected by the Central.
2) A method was added to the Central SDK observer protocol, allowing callback upon discovery of a device.
3) Internally, in the Central SDK, we have the "cascaded discovery" in place that discovers Peripherals, then Service for each peripheral, then Characteristics for each Service. The observer discovery callback is not made until all discoveries are complete.
4) The Peripheral Mode Screen has been removed from the WatchOS and TVOS app.
5) The Peripheral Mode SDK ha been reoved from the WatchOS and TVOS SDKs.

The reason for #4 and #5, is because Apple does not actually support running Peripheral Mode on Watches or TVs.

They don't make this very clear. It seems that the only place it is actually mentioned by them is in the 2017 WWDC Bluetooth session video.

What they do, is exclude the initializers for mutable Services and Characteristics from Watch and TV targets, so you can't actually build Peripherals on those platforms.